Various code snippets to unlock the Appwrite Document object data.
