+++ Rossi Notes/Models/Protocol.swift
@@ -7,19 +7,19 @@
 
 import Foundation
 
-struct Protocol: Codable, Identifiable {
-    let id: String
-    let name: String
-    let dogReactive: Bool
-    let barrierReactive: Bool
-    let miscNotes: String
-    let protocolDate: Date
-    let catReactive: Bool
-    let resourceGuarder: Bool
-    let strangerReactive: Bool
-    let jumpyMouthy: Bool
-    let doorRoutine: Bool
-    let placeRoutine: Bool
-    let leashReactive: Bool
-    let creatorName: String
-}
+//struct Protocol: Codable, Identifiable {
+//    let id: String
+//    let name: String
+//    let dogReactive: Bool
+//    let barrierReactive: Bool
+//    let miscNotes: String
+//    let protocolDate: Date
+//    let catReactive: Bool
+//    let resourceGuarder: Bool
+//    let strangerReactive: Bool
+//    let jumpyMouthy: Bool
+//    let doorRoutine: Bool
+//    let placeRoutine: Bool
+//    let leashReactive: Bool
+//    let creatorName: String
+//}
