+++ Rossi Notes/ViewModels/LoginViewModel.swift
@@ -6,9 +6,11 @@
 //
 
 import Foundation
+import Appwrite
 
 class LoginViewModel: ObservableObject {
-    @Published var email: String = ""
-    @Published var password: String = ""
+    @Published var email: String = "gboyd69@yahoo.com"
+    @Published var password: String = "11Gunner$"
     @Published var response: String = ""
+    @Published var session: Session?
 }
